2009-06-24  Brady Eidson  <beidson@apple.com>
        Fix 64-bit SnowLeopard build.

        * html/DataGridColumnList.cpp:
        (WebCore::DataGridColumnList::remove):
        (WebCore::DataGridColumnList::move):



Canonical link: https://commits.webkit.org/36859@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@45098 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/html/DataGridColumnList.cpp
@@ -57,7 +57,7 @@ PassRefPtr<DataGridColumn> DataGridColumnList::add(const String& id, const Strin
 
 void DataGridColumnList::remove(DataGridColumn* col)
 {
-    unsigned index = m_columns.find(col);
+    size_t index = m_columns.find(col);
     if (index == notFound)
         return;
     m_columns.remove(index);
@@ -69,7 +69,7 @@ void DataGridColumnList::remove(DataGridColumn* col)
 
 void DataGridColumnList::move(DataGridColumn* col, unsigned long index)
 {
-    unsigned colIndex = m_columns.find(col);
+    size_t colIndex = m_columns.find(col);
     if (colIndex == notFound)
         return;
     m_columns.insert(index, col);
